requirements.txt
discord.py
python-dotenv
langchain-core
langchain-groq
langchain-google-community
gspread
google-auth
google-api-python-client